break down of module_helper into smaller pieces, keeping compatibility
